The BRAVE BROWSER
For me, it's the BRAVE BROWSER. The best internet browser. I even download the BRAVE BROWSER on my phone, tablet, laptop, and desktop because the BRAVE BROWSER is so user friendly and helps fill my crypto wallets with BAT (Basic Attention Token)

One time I asked for 3 ads to be served and they gave me three, WITH the BAT (Basic Attention Token). I said, "Wow, three for free!" and the nice friendly BRAVE dev laughed and said, "I'm going to call you 3-for-free!". But the best part is with BAT, it's not even free for me!

Now the BRAVE BROWSER greets me with "hey it's 3-for-free!" and ALWAYS blocks every ad, cookie, and tracker. It's such a fun and cool atmosphere with my BRAVE, I use it at least 7 times a week for my daily browsing needs which include work, shitposting, school, and stalking my ex on facebook (without the trackers!).

What a great BRAVE BROWSER